+++ src/Service/Kinesis/src/Input/ListStreamConsumersInput.php
@@ -160,7 +160,7 @@ final class ListStreamConsumersInput extends Input
             $payload['MaxResults'] = $v;
         }
         if (null !== $v = $this->streamCreationTimestamp) {
-            $payload['StreamCreationTimestamp'] = $v->format(\DateTimeInterface::ATOM);
+            $payload['StreamCreationTimestamp'] = $v->getTimestamp();
         }
 
         return $payload;